% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hades.R
\docType{package}
\name{Hades-package}
\alias{Hades}
\alias{Hades-package}
\title{Hades: Health Analytics Data-to-Evidence Suite}
\description{
A collection of packages for executing data-analytics against the OMOP Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/Hades/}
  \item \url{https://github.com/OHDSI/Hades}
  \item Report bugs at \url{https://github.com/OHDSI/Hades/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
