% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Packages.R
\name{listHadesPackages}
\alias{listHadesPackages}
\title{List all packages currently in HADES}
\usage{
listHadesPackages()
}
\value{
A data frame with one row per HADES package.
}
\description{
List all packages currently in HADES
}
\examples{
listHadesPackages()

}
